python_tests/requirements.txt
pyserial>=3.5
typing-extensions>=4.0.0 